# swagger.yaml
openapi: 3.0.0
info:
  title: AWS Lambda Data Sync API
  version: 1.0.0
  description: API for uploading transaction files and retrieving aggregated transaction data in an e-commerce system.
servers:
  - url: http://127.0.0.1:8000
paths:
  /api/transactions/upload:
    post:
      summary: Upload Transaction File
      operationId: uploadTransactionFile
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '500':
          description: Error occurred during file upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
...